add is_schema_valid helper function
